Reorganize non-parents permission rules
Instead of creating a big fat class with lots of rules and complex methods, I'm going to maintain the current `PermissionRules` but renamed to `RelationshipRules` as these rules are specific to the relationship between applicant-child.

And a new class will be created for the "application-wide" permission rules that are needed towards the end of the journey.
